Changed mention of disabling s-vMotion to sDRS
sDRS allows automated svMotion to occur, otherwise it only occurs manually.

You wouldn’t want to actually disable svMotion (not even sure that you can), but you would want to prevent automated svMotions to occur in this case. So this fixes the wording on the warning.

I recommend this is pushed to all versions of the docs where this warning appears.